+++ src/app/services/application/application.service.ts
@@ -1,7 +1,5 @@
 import { Injectable } from '@angular/core';
 import { fromEvent, Observable, Subject } from 'rxjs';
-import { debounceTime } from 'rxjs/operators';
-import { Constants } from '../../common/application/constants';
 import { BaseApplicationService } from './base-application.service';
 
 @Injectable()
@@ -10,17 +8,13 @@ export class ApplicationService implements BaseApplicationService {
     private mouseButtonReleased: Subject<void> = new Subject();
 
     constructor() {
-        fromEvent(window, 'resize')
-            .pipe(debounceTime(Constants.albumsRedrawDelayMilliseconds))
-            .subscribe((event: any) => {
-                this.windowSizeChanged.next();
-            });
+        fromEvent(window, 'resize').subscribe((event: any) => {
+            this.windowSizeChanged.next();
+        });
 
-        fromEvent(document, 'mouseup')
-            .pipe(debounceTime(Constants.albumsRedrawDelayMilliseconds))
-            .subscribe((event: any) => {
-                this.mouseButtonReleased.next();
-            });
+        fromEvent(document, 'mouseup').subscribe((event: any) => {
+            this.mouseButtonReleased.next();
+        });
     }
 
     public windowSizeChanged$: Observable<void> = this.windowSizeChanged.asObservable();