+++ src/app/services/playback-information/base-playback-information.service.ts
@@ -1,0 +1,9 @@
+import { Observable } from 'rxjs';
+import { PlaybackInformation } from './playback-information';
+
+export abstract class BasePlaybackInformationService {
+    public abstract playingNextTrack$: Observable<PlaybackInformation>;
+    public abstract playingPreviousTrack$: Observable<PlaybackInformation>;
+    public abstract playingNoTrack$: Observable<PlaybackInformation>;
+    public abstract getCurrentPlaybackInformationAsync(): Promise<PlaybackInformation>;
+}